pyproject.toml
[tool.poetry]
name = "slack-emoji-uploader"
version = "0.0.1"
description = "Help to upload slack image"
authors = ["bearcc01205 <bearcc01205@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
backoff = "^2.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
